+++ Sadie.Networking.Writers/Players/Inventory/PlayerInventoryRefreshWriter.cs
@@ -1,5 +1,5 @@
-using Sadie.Shared.Networking;
-using Sadie.Shared.Networking.Packets;
+using Sadie.Shared.Unsorted.Networking;
+using Sadie.Shared.Unsorted.Networking.Packets;
 
 namespace Sadie.Networking.Writers.Players.Inventory;
 
@@ -7,6 +7,6 @@ public class PlayerInventoryRefreshWriter : NetworkPacketWriter
 {
     public PlayerInventoryRefreshWriter()
     {
-        WriteInteger(ServerPacketId.PlayerInventoryRefresh);
+        WriteShort(ServerPacketId.PlayerInventoryRefresh);
     }
 }
